Renamed filterSearch to filterUtil and fixed imports
